+++ creditapplication-backend/src/main/java/com/n11project/creditapplication/strategy/StrategyContext.java
@@ -18,15 +18,15 @@ public class StrategyContext {
 
     private final List<CalculateLimitStrategy> calculateLimitStrategyList;
 
-    public BigDecimal getCorrectStrategy(BigDecimal monthlyIncome , Integer creditScore, BigDecimal assurance){
+    public Double getCorrectStrategy(Double monthlyIncome , Integer creditScore, Double assurance){
         CalculateLimitStrategy calculateLimitStrategy = calculateLimitStrategyList.stream().filter(strategy -> strategy.isSuitable(monthlyIncome,creditScore)).findFirst().orElseThrow(StrategyNotFoundException::new);
         return calculateLimitStrategy.calculateLimit(monthlyIncome,assurance);
     }
 
-    public Application calculateLimitAndSetStatus(BigDecimal monthlyIncome , Integer creditScore , BigDecimal assurance){
+    public Application calculateLimitAndSetStatus(Double monthlyIncome , Integer creditScore , Double assurance){
         Application application = new Application();
         setApplicationStatus(creditScore, application);
-        BigDecimal creditLimit = getCorrectStrategy(monthlyIncome,creditScore,assurance);
+        Double creditLimit = getCorrectStrategy(monthlyIncome,creditScore,assurance);
         application.setCreditLimit(creditLimit);
         return application;
     }
